// src/types/IPCTypes.ts
export enum IPCTypes {
  AUTH_PASSWORD = 'AUTH_PASSWORD',
  AUTH_PASSWORD_RESPONSE = 'AUTH_PASSWORD_RESPONSE',
  AUTH_PASSWORD_RESPONSE_REGISTER = 'AUTH_PASSWORD_RESPONSE_REGISTER',
  AUTH_LOGIN = 'AUTH_LOGIN',
  AUTH_LOGIN_RESPONSE = 'AUTH_LOGIN_RESPONSE',
  VERIFY_DATABASE_PASSWORD = 'VERIFY_DATABASE_PASSWORD',
  VERIFY_DATABASE_PASSWORD_RESPONSE = 'VERIFY_DATABASE_PASSWORD_RESPONSE',
  GET_PRIVATE_KEY = 'GET_PRIVATE_KEY',
  GET_PRIVATE_KEY_RESPONSE = 'GET_PRIVATE_KEY_RESPONSE',
  GET_PUBLIC_KEY = 'GET_PUBLIC_KEY',
  GET_PUBLIC_KEY_RESPONSE_VALIDATE_USER = 'GET_PUBLIC_KEY_RESPONSE_VALIDATE_USER',
  GET_PUBLIC_KEY_RESPONSE = 'GET_PUBLIC_KEY_RESPONSE',
  UPDATE_DATABASE = 'UPDATE_DATABASE',
  UPDATE_DATABASE_RESPONSE = 'UPDATE_DATABASE_RESPONSE',
  REFRESH_ALL_ORGANIZATIONS = 'REFRESH_ALL_ORGANIZATIONS',
  REFRESH_ALL_ORGANIZATIONS_RESPONSE = 'REFRESH_ALL_ORGANIZATIONS_RESPONSE',
  REFRESH_ALL_ORGANIZATIONS_REFRESH_RESPONSE = 'REFRESH_ALL_ORGANIZATIONS_REFRESH_RESPONSE',
  SET_USER_CONFIG = 'SET_USER_CONFIG',
  SET_USER_CONFIG_RESPONSE = 'SET_USER_CONFIG_RESPONSE',
  INSERT_DATABASE_KEYS = 'INSERT_DATABASE_KEYS',
  INSERT_DATABASE_KEYS_RESPONSE = 'INSERT_DATABASE_KEYS_RESPONSE',
  GET_USER = 'GET_USER',
  GET_USER_RESPONSE = 'GET_USER_RESPONSE',
  LIST_SAFE_BOXES = 'LIST_SAFE_BOXES',
  LIST_SAFE_BOXES_RESPONSE = 'LIST_SAFE_BOXES_RESPONSE',
  UPDATE_USER_CONFIG = 'UPDATE_USER_CONFIG',
  UPDATE_USER_CONFIG_RESPONSE = 'UPDATE_USER_CONFIG_RESPONSE',
  LIST_MY_INVITES = 'LIST_MY_INVITES',
  LIST_MY_INVITES_RESPONSE = 'LIST_MY_INVITES_RESPONSE',
  CREATE_ORGANIZATION = 'CREATE_ORGANIZATION',
  CREATE_ORGANIZATION_RESPONSE = 'CREATE_ORGANIZATION_RESPONSE',
  CREATE_SAFE_BOX = 'CREATE_SAFE_BOX',
  CREATE_SAFE_BOX_RESPONSE = 'CREATE_SAFE_BOX_RESPONSE',
  EDIT_SAFE_BOX = 'EDIT_SAFE_BOX',
  EDIT_SAFE_BOX_RESPONSE = 'EDIT_SAFE_BOX_RESPONSE',
  DELETE_SAFE_BOX = 'DELETE_SAFE_BOX',
  DELETE_SAFE_BOX_RESPONSE = 'DELETE_SAFE_BOX_RESPONSE',
  DECRYPT_TEXT = 'DECRYPT_TEXT',
  DECRYPT_TEXT_RESPONSE = 'DECRYPT_TEXT_RESPONSE',
  UPDATE_SAFE_BOX = 'UPDATE_SAFE_BOX',
  UPDATE_SAFE_BOX_RESPONSE = 'UPDATE_SAFE_BOX_RESPONSE',
  DELETE_PRIVATE_KEY = 'DELETE_PRIVATE_KEY',
  DELETE_PRIVATE_KEY_RESPONSE = 'DELETE_PRIVATE_KEY_RESPONSE',
  ADD_NEW_PARTICIPANT_ORGANIZATION = 'ADD_NEW_PARTICIPANT_ORGANIZATION',
  ADD_NEW_PARTICIPANT_ORGANIZATION_RESPONSE = 'ADD_NEW_PARTICIPANT_ORGANIZATION_RESPONSE',
  DELETE_ORGANIZATION = 'DELETE_ORGANIZATION',
  DELETE_ORGANIZATION_RESPONSE = 'DELETE_ORGANIZATION_RESPONSE',
  UPDATE_ORGANIZATION = 'UPDATE_ORGANIZATION',
  UPDATE_ORGANIZATION_RESPONSE = 'UPDATE_ORGANIZATION_RESPONSE',
  REFRESH_TOKEN = 'REFRESH_TOKEN',
  REFRESH_TOKEN_RESPONSE = 'REFRESH_TOKEN_RESPONSE',
  CREATE_USER = 'CREATE_USER',
  CREATE_USER_RESPONSE = 'CREATE_USER_CONTROLLER',
  VERIFY_USER_REGISTERED = 'VERIFY_USER_REGISTERED',
  VERIFY_USER_REGISTERED_RESPONSE = 'VERIFY_USER_REGISTERED_RESPONSE',
  REMOVE_PARTICIPANT_ORGANIZATION = 'REMOVE_PARTICIPANT_ORGANIZATION',
  REMOVE_PARTICIPANT_ORGANIZATION_RESPONSE = 'REMOVE_PARTICIPANT_ORGANIZATION_RESPONSE',
  REMOVE_INVITE_PARTICIPANT = 'REMOVE_INVITE_PARTICIPANT',
  REMOVE_INVITE_PARTICIPANT_RESPONSE = 'REMOVE_INVITE_PARTICIPANT_RESPONSE',
  ADD_SAFE_BOX_USERS = 'ADD_SAFE_BOX_USERS',
  ADD_SAFE_BOX_USERS_RESPONSE = 'ADD_SAFE_BOX_USERS_RESPONSE',
  UPDATE_USERS_SAFE_BOX = 'UPDATE_USERS_SAFE_BOX',
  UPDATE_USERS_SAFE_BOX_RESPONSE = 'UPDATE_USERS_SAFE_BOX_RESPONSE',
  DECLINE_ORGANIZATION_INVITE = 'DECLINE_ORGANIZATION_INVITE',
  DECLINE_ORGANIZATION_INVITE_RESPONSE = 'DECLINE_ORGANIZATION_INVITE_RESPONSE',
  ACCEPT_ORGANIZATION_INVITE = 'ACCEPT_ORGANIZATION_INVITE',
  ACCEPT_ORGANIZATION_INVITE_RESPONSE = 'ACCEPT_ORGANIZATION_INVITE_RESPONSE',
  LEAVE_ORGANIZATION = 'LEAVE_ORGANIZATION',
  LEAVE_ORGANIZATION_RESPONSE = 'LEAVE_ORGANIZATION_RESPONSE',
  REFRESH_SAFE_BOXES = 'REFRESH_SAFE_BOXES',
  REFRESH_SAFE_BOXES_RESPONSE = 'REFRESH_SAFE_BOXES_RESPONSE',
  REFRESH_ALL_SAFE_BOXES = 'REFRESH_ALL_SAFE_BOXES',
  REFRESH_ALL_SAFE_BOXES_RESPONSE = 'REFRESH_ALL_SAFE_BOXES_RESPONSE',
  SESSION_EXPIRED = 'SESSION_EXPIRED',
  FORCE_REFRESH_SAFE_BOXES = 'FORCE_REFRESH_SAFE_BOXES',
  FORCE_REFRESH_SAFE_BOXES_RESPONSE = 'FORCE_REFRESH_SAFE_BOXES_RESPONSE',
  CHANGE_SAFETY_PHRASE = 'CHANGE_SAFETY_PHRASE',
  CHANGE_SAFETY_PHRASE_RESPONSE = 'CHANGE_SAFETY_PHRASE_RESPONSE',
  GENERATE_PAR_KEYS_RESPONSE = 'GENERATE_PAR_KEYS_RESPONSE',
}